[Chrome Next] Move app menu to kbn-ui
