+++ packages/kbn-rspack-optimizer/limits.yml
@@ -186,7 +186,7 @@ pageLoadAssetSize:
   sessionView: 379
   share: 379
   shared-core: 348361
-  shared-misc: 97669
+  shared-misc: 114567
   shared-packages: 4546017
   shared-plugins: 13536562
   shared-solution-packages: 347767
