+++ src/platform/kbn-ui/app-menu/index.ts
@@ -7,17 +7,33 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
+export { AppMenuComponent, type AppMenuBreakpointSource } from './src';
+export { AppMenuItem } from './src';
+export { AppMenuActionButton } from './src';
+export { AppMenuOverflowButton } from './src';
+export { AppMenuPopover } from './src';
+export { AppMenuPopoverActionButtons } from './src';
+
+export type {
+  AppMenuRunAction,
+  AppMenuRunActionParams,
+  AppMenuConfig,
+  AppMenuItemType,
+  AppMenuPrimaryActionItem,
+  AppMenuPopoverItem,
+  AppMenuSplitButtonProps,
+  AppMenuStaticItem,
+} from './src';
+
+export { APP_MENU_ITEM_LIMIT } from './src';
+
 export {
-  AppMenuComponent,
-  AppMenuItem,
-  AppMenuActionButton,
-  AppMenuOverflowButton,
-  AppMenuPopover,
-  AppMenuPopoverActionButtons,
-  APP_MENU_ITEM_LIMIT,
   APP_MENU_TEST_SUBJECTS,
   getAppMenuItemTestSubj,
   getAppMenuActionButtonTestSubj,
+} from './src';
+
+export {
   getDisplayedItemsAllowedAmount,
   getShouldOverflow,
   isDisabled,
@@ -28,16 +44,4 @@ export {
   getPopoverActionItems,
   getIsSelectedColor,
   hasNonGlobalStaticItems,
-} from '@kbn/ui-app-menu';
-
-export type {
-  AppMenuBreakpointSource,
-  AppMenuRunAction,
-  AppMenuRunActionParams,
-  AppMenuConfig,
-  AppMenuItemType,
-  AppMenuPrimaryActionItem,
-  AppMenuPopoverItem,
-  AppMenuSplitButtonProps,
-  AppMenuStaticItem,
-} from '@kbn/ui-app-menu';
+} from './src';
