+++ src/platform/kbn-ui/app-menu/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-browser",
+  "id": "@kbn/ui-app-menu",
+  "owner": "@elastic/appex-sharedux",
+  "group": "platform",
+  "visibility": "private"
+}
