+++ src/platform/kbn-ui/app-menu/moon.yml
@@ -1,9 +1,9 @@
 # This file is generated by the @kbn/moon package. Any manual edits will be erased!
 #  To extend this, write your extensions/overrides to 'moon.extend.yml'
-#  then regenerate this file with: 'node scripts/regenerate_moon_projects.js --update --filter @kbn/core-chrome-app-menu-components'
+#  then regenerate this file with: 'node scripts/regenerate_moon_projects.js --update --filter @kbn/ui-app-menu'
 
 $schema: https://moonrepo.dev/schemas/project.json
-id: '@kbn/core-chrome-app-menu-components'
+id: '@kbn/ui-app-menu'
 layer: unknown
 owners:
   defaultOwner: '@elastic/appex-sharedux'
@@ -11,21 +11,26 @@ toolchains:
   default: node
 language: typescript
 project:
-  title: '@kbn/core-chrome-app-menu-components'
-  description: Moon project for @kbn/core-chrome-app-menu-components
+  title: '@kbn/ui-app-menu'
+  description: Moon project for @kbn/ui-app-menu
   channel: ''
   owner: '@elastic/appex-sharedux'
-  sourceRoot: src/core/packages/chrome/app-menu/core-chrome-app-menu-components
+  sourceRoot: src/platform/kbn-ui/app-menu
 dependsOn:
-  - '@kbn/ui-app-menu'
+  - '@kbn/i18n'
+  - '@kbn/ui-chrome-layout-utils'
 tags:
   - shared-browser
   - package
   - prod
   - group-platform
-  - shared
+  - private
+  - jest-unit-tests
 fileGroups:
   src:
     - '**/*.ts'
+    - '**/*.tsx'
     - '!target/**/*'
+  jest-config:
+    - jest.config.js
 tasks: {}
