+++ src/platform/kbn-ui/app-menu/setup_tests.ts
