+++ src/platform/kbn-ui/app-menu/src/components/app_menu_badge.test.tsx
