+++ src/platform/kbn-ui/app-menu/src/constants.ts
