+++ src/platform/kbn-ui/app-menu/src/utils.tsx
