+++ tsconfig.base.json
@@ -2723,6 +2723,8 @@
       "@kbn/ui-actions-browser/*": ["./src/platform/packages/shared/kbn-ui-actions-browser/*"],
       "@kbn/ui-actions-plugin": ["./src/platform/plugins/shared/ui_actions"],
       "@kbn/ui-actions-plugin/*": ["./src/platform/plugins/shared/ui_actions/*"],
+      "@kbn/ui-app-menu": ["./src/platform/kbn-ui/app-menu"],
+      "@kbn/ui-app-menu/*": ["./src/platform/kbn-ui/app-menu/*"],
       "@kbn/ui-callout": ["./src/platform/kbn-ui/callout"],
       "@kbn/ui-callout/*": ["./src/platform/kbn-ui/callout/*"],
       "@kbn/ui-chrome-layout": ["./src/platform/kbn-ui/chrome-layout"],
