+++ x-pack/platform/plugins/private/translations/translations/de-DE.json
@@ -1108,8 +1108,6 @@
     "core.chrome.appHeader.addIntegrationsMenuItemLabel": "Integrationen hinzufügen",
     "core.chrome.appHeader.documentationMenuItemLabel": "Dokumentation",
     "core.chrome.appHeader.feedbackMenuItemLabel": "Feedback",
-    "core.chrome.appMenu.showMoreButtonLabel": "Mehr",
-    "core.chrome.appMenu.showMoreButtonTooltip": "Mehr",
     "core.chrome.legacyBrowserWarning": "Ihr Browser erfüllt nicht die Sicherheitsanforderungen für Kibana.",
     "core.deprecations.apiAccessDeprecation.apiCallsDetailsMessage": "Die API „{routeWithMethod}“ wurde {apiTotalCalls} Mal aufgerufen. Der letzte Aufruf ist am {apiLastCalledAt} erfolgt.",
     "core.deprecations.apiAccessDeprecation.enableDebugLogsMessage": "Um Informationen über Aufrufe von Elastic-internen APIs in Debug-Protokolle aufzunehmen, bearbeiten Sie Ihre Kibana-Konfiguration wie in [der Dokumentation] ({enableDeprecationHttpDebugLogsLink}) beschrieben.",
