+++ x-pack/platform/plugins/private/translations/translations/fr-FR.json
@@ -1107,8 +1107,6 @@
     "core.chrome.appHeader.addIntegrationsMenuItemLabel": "Ajouter des intégrations",
     "core.chrome.appHeader.documentationMenuItemLabel": "Documentation",
     "core.chrome.appHeader.feedbackMenuItemLabel": "Commentaires",
-    "core.chrome.appMenu.showMoreButtonLabel": "Plus",
-    "core.chrome.appMenu.showMoreButtonTooltip": "Plus",
     "core.chrome.legacyBrowserWarning": "Votre navigateur ne satisfait pas aux exigences de sécurité de Kibana.",
     "core.deprecations.apiAccessDeprecation.apiCallsDetailsMessage": "L'API « {routeWithMethod} » a été appelée {apiTotalCalls} fois. Le dernier appel est survenu le {apiLastCalledAt}.",
     "core.deprecations.apiAccessDeprecation.enableDebugLogsMessage": "Pour inclure des informations dans les logs de débogage au sujet des appels aux API internes à Elastic, modifiez votre configuration Kibana comme décrit dans [la documentation]({enableDeprecationHttpDebugLogsLink}).",
