+++ x-pack/platform/plugins/private/translations/translations/ja-JP.json
@@ -1109,8 +1109,6 @@
     "core.chrome.appHeader.addIntegrationsMenuItemLabel": "統合の追加",
     "core.chrome.appHeader.documentationMenuItemLabel": "ドキュメント",
     "core.chrome.appHeader.feedbackMenuItemLabel": "フィードバック",
-    "core.chrome.appMenu.showMoreButtonLabel": "詳細",
-    "core.chrome.appMenu.showMoreButtonTooltip": "詳細",
     "core.chrome.legacyBrowserWarning": "ご使用のブラウザが Kibana のセキュリティ要件を満たしていません。",
     "core.deprecations.apiAccessDeprecation.apiCallsDetailsMessage": "API \"{routeWithMethod}\"が{apiTotalCalls}回呼び出されました。前回の呼び出しは{apiLastCalledAt}でした。",
     "core.deprecations.apiAccessDeprecation.enableDebugLogsMessage": "Elastic内部のAPI呼び出しに関する情報をデバッグログに含めるには、[ドキュメント]({enableDeprecationHttpDebugLogsLink})の詳細に従ってKibana構成を編集してください。",
