+++ x-pack/platform/plugins/private/translations/translations/zh-CN.json
@@ -1109,8 +1109,6 @@
     "core.chrome.appHeader.addIntegrationsMenuItemLabel": "添加集成",
     "core.chrome.appHeader.documentationMenuItemLabel": "文档",
     "core.chrome.appHeader.feedbackMenuItemLabel": "反馈",
-    "core.chrome.appMenu.showMoreButtonLabel": "更多",
-    "core.chrome.appMenu.showMoreButtonTooltip": "更多",
     "core.chrome.legacyBrowserWarning": "您的浏览器不满足 Kibana 的安全要求。",
     "core.deprecations.apiAccessDeprecation.apiCallsDetailsMessage": "API“{routeWithMethod}”已被调用 {apiTotalCalls} 次。上一次调用时间为 {apiLastCalledAt}。",
     "core.deprecations.apiAccessDeprecation.enableDebugLogsMessage": "要在调试日志中包括有关 Elastic 内部的 API 调用的信息，请按照[文档]({enableDeprecationHttpDebugLogsLink}) 中的详细说明编辑 Kibana 配置。",
